requirements.txt
PyMysql==0.9.3
djangorestframework-jwt==1.11.0
djangorestframework==3.9.1
markdown==3.0.1
django-filter==2.1.0
django-crispy-forms==1.7.2
django-reversion==3.0.2
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
six==1.10.0
requests==2.21.0
coreapi==2.3.3
django-cors-headers==2.4.0
xlwt==1.3.0
xlsxwriter==1.1.2
drf-extensions
django-redis
pycryptodome
raven